@W-19688128 Order Details slight refactors and fixes
