+++ packages/template-retail-react-app/app/utils/shipment-utils.js
@@ -193,3 +193,24 @@ export const isDefaultShipmentEmpty = (basket) => {
 
     return !basket.productItems?.some((item) => item.shipmentId === DEFAULT_SHIPMENT_ID)
 }
+
+/**
+ * Groups shipments into pickup and delivery arrays
+ * @param {Object} order - The order or basket object containing shipments
+ * @returns {Object} Object with pickup and delivery arrays of shipments
+ */
+export const groupShipmentsByDeliveryOption = (order) => {
+    const pickupShipments = []
+    const deliveryShipments = []
+
+    order?.shipments?.forEach((shipment) => {
+        const isPickup = isPickupShipment(shipment)
+        if (isPickup) {
+            pickupShipments.push(shipment)
+        } else {
+            deliveryShipments.push(shipment)
+        }
+    })
+
+    return {pickupShipments, deliveryShipments}
+}
